requirements.txt
asgiref==3.2.10
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-registration==3.1
django-rest-auth==0.9.5
django-webpack-loader==0.7.0
djangorestframework==3.11.1
docopt==0.6.2
greenlet==0.4.16
gunicorn==20.0.4
idna==2.10
msgpack==1.0.0
neovim==0.3.1
oauthlib==3.1.0
Pillow==7.2.0
psutil==5.7.2
pynvim==0.4.2
python-xlib==0.28
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
ueberzug==18.1.6
urllib3==1.25.10
whitenoise==5.2.0
